% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{dis.data.re}
\alias{dis.data.re}
\title{Line transect data with redistribution post-impact}
\format{A data frame with 10951 rows and 12 variables}
\description{
A simulated dataset containing the observed perpendicular distances, the effort data and other variables of
segmented line transect data. The variables are as follows:
}
\details{
\itemize{
\item \code{transect.id} Identifier for the individual visits to the transects
\item \code{transect.label} Labels for transects
\item \code{season} Numerical indicator for the four different seasons
\item \code{impact} Numerical indicator for before (0) and after (1) impact
\item \code{segment.id} Identifier for individual visits to the segment
\item \code{segment.label} Label for segments
\item \code{length} Length of segment in km
\item \code{x.pos} spatial location in the horizontal axis in UTMs
\item \code{y.pos} spatial location in the vertical axis in UTMs
\item \code{depth} Depth in m
\item \code{object} Id for detected object
\item \code{distance} Perpendicular distance from the line
}
}
\keyword{datasets}

