% Generated by roxygen2 (4.0.1): do not edit by hand
\name{exchange.step_2d}
\alias{exchange.step_2d}
\title{Function for exchanging knot locations and re-fitting model to find best one}
\usage{
exchange.step_2d(gap, knotDist, radii, invInd, dists, explData, response,
  explanatory, maxIterations, fitnessMeasure, point, knotPoint, position, aR,
  BIC, track, out.lm, improveEx, maxKnots, tol = 0, baseModel, radiusIndices,
  models, interactionTerm, data)
}
\description{
Function for exchanging knot locations and re-fitting model to find best one
}
\author{
Cameron Walker, Department of Enginering Science, University of Auckland.
}

