% Generated by roxygen2 (4.0.1): do not edit by hand
\name{improve.step_2d}
\alias{improve.step_2d}
\title{Function to move knots to neighbours to see if there is any improvement in fit}
\usage{
improve.step_2d(gap, knotDist, radii, invInd, dists, gridResp, grid, explData,
  xVals, yVals, num, response, explanatory, maxIterations, fitnessMeasure,
  point, knotPoint, position, aR, BIC, track, out.lm, improveNudge, tol = 0,
  baseModel, radiusIndices, models, interactionTerm, data)
}
\description{
Function to move knots to neighbours to see if there is any improvement in fit
}
\author{
Cameron Walker, Department of Enginering Science, University of Auckland.
}

