% Generated by roxygen2 (4.0.1): do not edit by hand
\name{thinModels}
\alias{thinModels}
\title{function to thin the number of models}
\usage{
thinModels(models)
}
\description{
function to thin the number of models
}
\author{
Cameron Walker, Department of Enginering Science, University of Auckland.
}

